+++ packages/flutter_scene/lib/scene.dart
@@ -79,6 +79,7 @@ export 'src/importer/scene_registry.dart'
 
 export 'src/ambient_occlusion.dart'
     show AmbientOcclusionSettings, SpecularAmbientOcclusionMode;
+export 'src/depth_of_field.dart' show DepthOfField, DepthOfFieldQuality;
 export 'src/fog.dart' show Fog, FogMode;
 export 'src/god_rays.dart' show GodRaysSettings;
 export 'src/screen_space_reflections.dart'